Refactor deployments to have names
This patch allows a deployment to have multiple sub-instances
of a same sub-deployment but with different names. Technically
this is done by prepending the deployment name to the resource names
inside terraform.

This is a breaking change!

- output function callback will now have a new parameter
"deployment_name" as the first parameter.

- all resource references should now use id_of() or output_of()
macros. Example: output_of("aws_instance", "master", "public_ip")